Text editor outlook (#28)
* yarn add codemirror, markdown-it, highlightjs

codemirror -- 將輸入區的 md 語法 highlight

markdown-it -- 將輸入區的 md code 轉換成 html

highlightjs -- 將 md code 中的程式碼在轉換成 html 時highlight

* codemirror part

1- import codemirror in webpack

2- 完成 codemirror 基本設定

* import 語言檢查的 js 程式

1- import codemirror 黑色主題

2- import 我們熟悉的語言 highlihgt js script

* import markdown-it

1- 設定好 markdown-it

2- 模擬即時顯示

* import highlightjs

1- md to html 裡面的程式碼能正常 highlight

2- code highlight 主題是 github

* 調整 text_editor

1- css 調整，輸入區與顯示區左右對半

2- 將 text_editor 的 html 抽出成一個檔案， render 給 editor_new and editor_edit

* 增加編輯書本內容的連結

* 修正編輯書本內容的連結與解決左側輸入區無法顯示

* 當左側輸入區有變化才渲染內容

* 解決md to html 圖片顯示故障藉由 調整 md to html 渲染時間

* 左側文字輸入區與右側畫面渲染區左右皆可同步滑動

* tailwind 測試

1- 會 reset css ，md to html 的渲染結果也會 reset

2- text_editor 改成scss 之後方便把 md to html 的設定補上

* 修好因為tailwind reset css 而故障的 md to html 顯示區
