--- app/javascript/stylesheets/text_editor/text_editor.css
@@ -1,20 +1,0 @@
-.wrap {
-  display: flex;
-  width: 100%;
-}
-.custom {
-  display: block;
-  width: 50%;
-  height: 100vh;
-}
-.custom .CodeMirror {
-  width: 100%;
-  height: 100vh;
-  font-size: 20px;
-}
-#targetDiv {
-  border: 1px dashed #333333;
-  width: 50%;
-  height: 100vh;
-  padding: 0px 5px 0px 5px;
-}
